+++ src/lib/components/TimePicker/components/TimePickerContainer.tsx
@@ -1,53 +1,25 @@
 import { useCallback, useContext } from "react";
-import TimeLabel from "@/components/TimePicker/components/TimeLabel";
-import TimeSelector from "@/components/TimePicker/components/TimeSelector/TimeSelector";
-import TimePeriodSelector from "@/components/TimePicker/components/TimePeriodSelector";
-import { TimePickerContext, TimePickerProvider } from "@/components/TimePicker/context/TimePickerProvider";
-import { Time, TimePickerProviderProps } from "@/components/TimePicker/types";
+import TimePickerContent from "@/components/TimePicker/components/TimePickerContent";
+import { TimePickerContext } from "@/components/TimePicker/context/TimePickerProvider";
+import { Time } from "@/components/TimePicker/types";
 import PickerActions from "@/components/Motif/Pickers/components/PickerActions";
-import styles from "../TimePicker.module.scss";
-import { StandardProps } from "../../../types";
 
-type Props = TimePickerProviderProps & {
+type Props = {
   onOkClick?: (time?: Time) => void;
   removeActionButtons?: boolean;
-  removeLabel?: boolean;
-} & StandardProps;
+};
 
-type ContentProps = {
-  onOkClick?: (time?: Time) => void;
-  removeActionButtons?: boolean;
-  removeLabel?: boolean;
-} & StandardProps;
-
-const TimePickerContainerContent = (props: ContentProps) => {
-  const { onOkClick, removeLabel, removeActionButtons, className } = props;
-  const { size, resetTime, time, timePeriod } = useContext(TimePickerContext);
+const TimePickerContainer = (props: Props) => {
+  const { onOkClick, removeActionButtons } = props;
+  const { size, resetTime, time } = useContext(TimePickerContext);
 
   const okClickHandler = useCallback(() => onOkClick?.(time), [onOkClick, time]);
 
   return (
-    <div className={`${styles[size]} ${className ?? ""}`.trim()}>
-      {!removeLabel && <TimeLabel />}
-      <TimeSelector />
-      {timePeriod && <TimePeriodSelector />}
+    <>
+      <TimePickerContent />
       {!removeActionButtons && <PickerActions size={size} onOkClick={okClickHandler} onClearClick={resetTime} spread />}
-    </div>
-  );
-};
-
-const TimePickerContainer = (props: Props) => {
-  const { onOkClick, removeLabel, removeActionButtons, className, ...providerProps } = props;
-
-  return (
-    <TimePickerProvider {...providerProps}>
-      <TimePickerContainerContent
-        onOkClick={onOkClick}
-        removeLabel={removeLabel}
-        removeActionButtons={removeActionButtons}
-        className={className}
-      />
-    </TimePickerProvider>
+    </>
   );
 };
